FINAL PUSH

+++ src/main/java/com/csa/samplefullstack/Entity/Employee.java
@@ -19,12 +19,12 @@ public class Employee {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employeeid;
+    private Long employeeid;
     private String name;
     private String department;
     private String location;
 
-    public Employee(int employeeid, String name, String department, String location) {
+    public Employee(Long employeeid, String name, String department, String location) {
         this.employeeid = employeeid;
         this.name = name;
         this.department = department;
@@ -36,7 +36,7 @@ public class Employee {
         //when initializing w/o parameters
     }
 
-    public int getEmployeeid() {
+    public Long getEmployeeid() {
         return employeeid;
     }
 
@@ -52,7 +52,7 @@ public class Employee {
         return location;
     }
 
-    public void setEmployeeid(int employeeid) {
+    public void setEmployeeid(Long employeeid) {
         this.employeeid = employeeid;
     }
